관리자-페이지-Admin-권한-설정 * 이제 관리자 id, pw로 로그인할 수 있다. * 정상 로그인 시 관리자 권한의 기능 사용가능

+++ src/main/java/com/ssafy/mugit/user/service/UserDeleteService.java
@@ -1,5 +1,7 @@
 package com.ssafy.mugit.user.service;
 
+import com.ssafy.mugit.global.exception.UserApiException;
+import com.ssafy.mugit.global.exception.error.UserApiError;
 import com.ssafy.mugit.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -10,6 +12,10 @@ public class UserDeleteService {
     private final UserRepository userRepository;
 
     public void deleteUserEntity(Long userId) {
+
+        // pk 검증
+        if (userId == null) throw new UserApiException(UserApiError.NOT_AUTHORIZED_USER);
+
         userRepository.deleteById(userId);
     }
 }
